refactor: move delete account to settings page and match padding
- Move delete account button and confirmation modal from profile to settings page
- Match settings page container padding/maxWidth to profile page (900px, 40/60px padding)
- Use warm white background (#FAFAF7) consistently across both pages
- Clean up unused imports from profile page
